# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Lead Generation System",
    "description": "Autonomous system to discover, qualify, enrich, and outreach to local businesses using AI-driven multi-touchpoint campaigns.",
    "version": "2.1.0"
  },
  "paths": {
...
    "/api/v1/reports": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "List Reports",
        "operationId": "list_reports_api_v1_reports_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReportResponse"
                  },
                  "type": "array",
                  "title": "Response List Reports Api V1 Reports Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "description": "Retrieves all daily pipeline reports sorted by date (newest first)."
      }
    },
    "/api/v1/reports/{date}": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Get Report By Date",
        "operationId": "get_report_by_date_api_v1_reports__date__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "description": "Retrieves the daily pipeline report for a specific calendar date."
      }
    },
    "/api/v1/reports/{date}/download": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Download Report",
        "operationId": "download_report_api_v1_reports__date__download_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "description": "Downloads the Excel workbook generated for the specified daily report."
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReportResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "report_date": {
            "type": "string",
            "format": "date",
            "title": "Report Date"
          },
          "leads_discovered": {
            "type": "integer",
            "title": "Leads Discovered"
          },
          "leads_qualified": {
            "type": "integer",
            "title": "Leads Qualified"
          },
          "emails_sent": {
            "type": "integer",
            "title": "Emails Sent"
          },
          "emails_opened": {
            "type": "integer",
            "title": "Emails Opened"
          },
          "links_clicked": {
            "type": "integer",
            "title": "Links Clicked"
          },
          "replies_received": {
            "type": "integer",
            "title": "Replies Received"
          },
          "new_conversions": {
            "type": "integer",
            "title": "New Conversions"
          },
          "pipeline_status": {
            "type": "string",
            "title": "Pipeline Status"
          },
          "pipeline_started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pipeline Started At"
          },
          "pipeline_ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pipeline Ended At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "report_date",
          "leads_discovered",
          "leads_qualified",
          "emails_sent",
          "emails_opened",
          "links_clicked",
          "replies_received",
          "new_conversions",
          "pipeline_status"
        ],
        "title": "ReportResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}